channels are unsubscribed on select button
